lint files. Move database and split routes into own files
