+++ server/routes/index.js
@@ -1,0 +1,13 @@
+const login = require('./login');
+const requestProfile = require('./request-profile');
+const signup = require('./signup');
+const updateProfile = require('./update-profile');
+const verify = require('./verify');
+
+module.exports = {
+  login,
+  requestProfile,
+  signup,
+  updateProfile,
+  verify
+};
